test(cli): go test modularity via build tags
Adds go build tags to integration tests
Specifies build tags in makefile
Disables policy testing

ALLY-603
