+++ integration/account_test.go
@@ -1,4 +1,5 @@
-//
+// +build account
+
 // Author:: Salim Afiune Maya (<afiune@lacework.net>)
 // Copyright:: Copyright 2021, Lacework Inc.
 // License:: Apache License, Version 2.0
