+++ integration/configure_test.go
@@ -1,4 +1,5 @@
-//
+// +build configure
+
 // Author:: Salim Afiune Maya (<afiune@lacework.net>)
 // Copyright:: Copyright 2020, Lacework Inc.
 // License:: Apache License, Version 2.0
