+++ integration/configure_unix_test.go
@@ -1,5 +1,5 @@
-// +build !windows
-//
+// +build configure configure,!windows
+
 // Author:: Salim Afiune Maya (<afiune@lacework.net>)
 // Copyright:: Copyright 2020, Lacework Inc.
 // License:: Apache License, Version 2.0
