+++ integration/configure_windows_test.go
@@ -1,5 +1,5 @@
-// +build !linux
-//
+// +build configure configure,!linux
+
 // Author:: Salim Afiune Maya (<afiune@lacework.net>)
 // Copyright:: Copyright 2020, Lacework Inc.
 // License:: Apache License, Version 2.0
